+++ src/App.js
@@ -1,5 +1,5 @@
 import BestTab from "./component/bestTab/BestTab.jsx";
-import Header from "./component/Header.jsx";
+import Header from "./component/header/Header.jsx";
 import ShowMoreBtn from "./component/ShowMoreBtn.jsx";
 import SlideContainer from "./component/slideContainer/SlideContainer.jsx";
 import GlobalStyle from "./style.js";
